requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-debug-toolbar==4.4.2
djangorestframework==3.15.2
iniconfig==2.0.0
model-bakery==1.18.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
